move headers to subdirectory, reference by relative path from test project.  delete UHCore merged file from test and UHCore projects.  remove old base.cpp/h files.

+++ CppInterface/robot.cpp
@@ -1,4 +1,4 @@
-#include "robot.h"
+#include "inc/robot.h"
 #include "boost/python.hpp"
 #include <string>
 #include <vector>
@@ -77,7 +77,7 @@ std::string Robot::setComponentState(std::string name,
 		std::vector<double> jointGoals, bool blocking) {
 
 	std::string format = "(s, [";
-	for (int i = 0; i < jointGoals.size(); i++) {
+	for (int i = 0; i < (int)jointGoals.size(); i++) {
 		format += "i,";
 	}
 	format = format.substr(0, format.length() - 1) + "], b)";
